make router work. wrap movie divs in links. style links. add fonts. restructure css. flex stuff
