--- app/assets/styles/main.css
@@ -1,7 +1,0 @@
-/*#main h1 {
-  background-color: red;
-}
-
-#main {
-  text-align: center;
-}*/
