+++ app/components/App/app.css
@@ -1,0 +1,13 @@
+
+.movie-watcher-h1 {
+  text-decoration: none;
+  font-family: 'Amatic SC', cursive;
+  font-weight: bold;
+  text-align: center;
+  color: red;
+}
+
+.movie-watcher-span {
+color: white;
+}
+
