+++ app/components/App/app.jsx
@@ -2,14 +2,12 @@ import React, { Component } from 'react';
 import MovieGridContainer from '../MovieGrid/MovieGridContainer';
 import MovieDetailContainer from '../MovieDetail/MovieDetailContainer';
 import { Route } from 'react-router-dom';
-import {Link} from 'react-router-dom';
-
-
+import { Link } from 'react-router-dom';
+import './app.css'
 
 export default class App extends Component {
   constructor() {
     super();
-
   }
 
   componentDidMount() {
@@ -26,9 +24,11 @@ export default class App extends Component {
 
     return (
       <div>
-        <Link to='/'><h1>Movie Watcher</h1></Link>
-        <Route exact path='/' component={MovieGridContainer} />
-        <Route exact path='/movie-detail' component={MovieDetailContainer} />
+        <Link className='movie-watcher-h1' to='/'><h1>Movie <span className='movie-watcher-span'>Watcher</span></h1></Link>
+        <Route exact path='/' component={ MovieGridContainer } />
+        <Route exact path='/movie-detail' render={({ match }) =>
+          <MovieDetailContainer />
+        } />
       </div>
     )
   }
@@ -36,4 +36,4 @@ export default class App extends Component {
 
 
 
-// <Route exact path='/movie-grid' component={MovieGridContainer} />
+
