+++ app/components/Movie/movie.css
@@ -1,7 +1,11 @@
+body {
+  background-color: black;
+}
+
 .movie-card{
   background-repeat: no-repeat;
   background-size: cover;
-  width: 22%;
+  width: 100%;
   max-width: 120px;
   height: 177px;
   margin: 5px;
@@ -10,3 +14,4 @@
 .movie-card:hover {
   cursor: pointer;
 }
+
