+++ index.html
@@ -3,6 +3,8 @@
   <head>
     <meta charset="utf-8">
     <title>MeatMeet</title>
+    <link href="https://fonts.googleapis.com/css?family=Oswald:400,500,600,700" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Amatic+SC:400,700" rel="stylesheet">
   </head>
   <body>
     <div id='main'></div>
