+++ package.json
@@ -20,7 +20,7 @@
     "pg-promise": "^3.2.3",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
-    "react-redux": "^5.0.2",
+    "react-redux": "^5.0.3",
     "react-router": "^3.0.1",
     "react-router-dom": "^4.0.0",
     "react-router-redux": "^5.0.0-alpha.4",
